[FlowInsight] fix physical view and insight head
